requirements.txt
numpy
opencv-python
rembg
Pillow
requests
selenium
